Update MAGGICcalc.js
The Diagnosis Date score contribution has been updated to accept both numeric and date-string based inputs. This is to account for the transformations made to the dataset necessary for retrospective analysis.

This change does not affect the source code of the KO's ability to process date based entry for prospective use of the KO in other settings.

The service. yaml file should be double checked to ensure that the the change of possible inputs for inputs.DateDiagnosed are still functional. Before the type of input was labeled as a "string" in the service. yaml. Now the string or numeric inputs can be accepted.